In order to avoid conflicts with other libraries, assigning $ with the jQuery namespace.
The coffescripts were compiling with a global use of the $ shortcut which can conflict with other libraries that could be included in an application.  By assigning the $ = jQuery, a var will be defined scoped to the function generated by compiling the coffee scripts.  This will allow for the functions to use $ internally as jQuery.

For example, if prototype is included, this avoids the two following errors on initialization of the page:

  editable_form.js?body=1:5 Uncaught TypeError: Cannot read property 'editableform' of undefined
      at editable_form.js?body=1:5
      at editable_form.js?body=1:55

and

  error_handling.js?body=1:2 Uncaught TypeError: Cannot read property 'editable' of undefined
      at error_handling.js?body=1:2
      at error_handling.js?body=1:9